skillful_agent/pyproject.toml
[project]
name = "skillful-agent"
version = "0.1.0"
description = "Natural language to PowerShell command agent"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.50",
    "google-adk>=1.0.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.0",
]

[project.scripts]
skillful-agent = "skillful_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
